+++ tests/test_core/test_files.py
@@ -33,18 +33,6 @@ class Test(TestCase):
     def test_get_log_path(self, _):
         self.assertEqual('home/.logsmith/app.log', files.get_log_path())
 
-    @mock.patch('app.core.files.Path.home', return_value='home')
-    def test_get_active_group_file_path(self, _):
-        self.assertEqual('home/.aws/active_group', files.get_active_group_file_path())
-
-    @mock.patch('app.core.files.Path.home', return_value='home')
-    def test_get_team_file_path(self, _):
-        self.assertEqual('home/.aws/active_team', files.get_team_file_path())
-
-    @mock.patch('app.core.files.Path.home', return_value='home')
-    def test_get_account_file_path(self, _):
-        self.assertEqual('home/.aws/active_account', files.get_account_file_path())
-
     def test__parse_yaml(self):
         yaml = files._parse_yaml('test: true')
         expected = {'test': True}